+++ packages/vite-app-ts/src/components/common/index.ts
@@ -1,2 +1,2 @@
 export * from './Ramp';
-export * from './ThemeSwitcher';
+export * from '../theme-switcher';
